requirements.txt
opencv-python
gradio>=4
numpy
Pillow
